Form Validation
This is a form validation in which two inputs are filled, name in the first input and password in the second. 
There is also a submit button and if it is filled without value then click on submit button. If selected it will show error or name will be filled. If value is less than three it will show error or in